# docs/swagger.yaml
definitions:
...
  service.AddToCartPayload:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
  service.AddressRes:
    properties:
      address:
        type: string
      country:
        type: string
      phoneNo:
        type: string
      state:
        type: string
      zipcode:
        type: string
    required:
    - address
    - country
    - phoneNo
    - state
    - zipcode
    type: object
  service.AllOrdersResponse:
    properties:
      address:
        $ref: '#/definitions/service.AddressRes'
      order:
        $ref: '#/definitions/service.OrderRes'
      product:
        $ref: '#/definitions/service.ProductRes'
    type: object
...
  service.GetCartResponse:
    properties:
      availability:
        type: string
      category:
        type: string
      createdAt:
        type: string
      depth:
        type: number
      description:
        type: string
      height:
        type: number
      id:
        type: integer
      image:
        type: string
      price:
        type: number
      quantity:
        type: integer
      rating:
        type: number
      returnPolicy:
        type: string
      shipping:
        type: string
      stock:
        type: integer
      thumbnail:
        type: string
      title:
        type: string
      updatedAt:
        type: string
      warranty:
        type: string
      weight:
        type: integer
      width:
        type: number
    type: object
...
  service.OrderRes:
    properties:
      orderId:
        type: string
      orderTime:
        type: string
      paymentMethod:
        type: string
      paymentStatus:
        type: string
      paymentTime:
        type: string
      quantity:
        type: number
    type: object
  service.OrdersPayload:
    properties:
      addressUsed:
        type: string
      paymentMethod:
        enum:
        - ONLINE
        - COD
        type: string
      products:
        items:
          $ref: '#/definitions/service.Product'
        type: array
    required:
    - addressUsed
    - paymentMethod
    type: object
  service.Product:
    properties:
      productId:
        type: string
      quantity:
        minimum: 1
        type: integer
    required:
    - productId
    - quantity
    type: object
  service.ProductRes:
    properties:
      category:
        type: string
      description:
        type: string
      image:
        type: string
      price:
        type: number
      productId:
        type: string
      thumbnail:
        type: string
      title:
        type: string
    type: object
...
  utils.ResUser:
    properties:
      message:
        type: string
    type: object
  utils.ResUserWithData:
    properties:
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: prachurjyagogoi123@gmail.com
  description: This is a dummy backend for an ecommerce store.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Commerce Backend API
  version: "1.0"
paths:
  /cart:
    get:
      description: API for fetching all the cart products.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.GetCartResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get all cart products.
      tags:
      - Cart
  /cart/product/{productId}:
    delete:
      description: API for removing a product from cart.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Remove product from cart.
      tags:
      - Cart
    post:
      consumes:
      - application/json
      description: API for adding an product the cart.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Cart payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/service.AddToCartPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Add product to cart.
      tags:
      - Cart
    put:
      consumes:
      - application/json
      description: API for updating the quantity for a cart product.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Product ID
        in: path
        name: productId
        required: true
        type: string
      - description: Cart payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/service.AddToCartPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Update a cart product.
      tags:
      - Cart
...
  /order:
    post:
      consumes:
      - application/json
      description: API for placing an order with the products in the cart.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order payload
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/service.OrdersPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Place an order.
      tags:
      - Order
  /orders:
    get:
      description: API for fetching all the orders of a user.
      parameters:
      - description: Bearer accessToken
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.ResUserWithData'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.AllOrdersResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResUser'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ResUser'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ResUser'
      summary: Get all orders of a user.
      tags:
      - Order
...
swagger: "2.0"